Update Tournament model and component
The Tournament model has been upgraded from an interface to a class to better match Angular's paradigm. On the tournaments component, a form has been added for searching for tournaments by name, status, and category. This allows for more flexible searching capabilities within the application. Additionally, the rendering for the tournament list in the tournaments.component.html has been simplified.